feat: add validation and custom exception

+++ TasksManagementApi.Application/UseCases/Tasks/Delete/DeleteTaskUseCase.cs
@@ -1,4 +1,5 @@
-﻿using TasksManagementApi.Infrastructure.Repositories.Interfaces;
+﻿using TasksManagementApi.Communication.Exceptions;
+using TasksManagementApi.Infrastructure.Repositories.Interfaces;
 
 namespace TasksManagementApi.Application.UseCases.Tasks.Delete;
 
@@ -12,7 +13,12 @@ public class DeleteTaskUseCase
     }
 
     public void Execute(Guid id)
-    { 
+    {
+        var task = _taskRepository.GetById(id);
+
+        if (task is null)
+            throw new NotFoundException("Tarefa não encontrada");
+
         _taskRepository.Delete(id);
     }
 }
